updated the Readme
