+++ javav2/example_code/secrets-manager/src/main/java/com/example/secrets/GetSecretValue.java
@@ -1,0 +1,72 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.example.secrets;
+
+// snippet-start:[secretsmanager.java2.get_secret.main]
+// snippet-start:[secretsmanager.java2.get_secret.import]
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.secretsmanager.SecretsManagerClient;
+import software.amazon.awssdk.services.secretsmanager.model.GetSecretValueRequest;
+import software.amazon.awssdk.services.secretsmanager.model.GetSecretValueResponse;
+import software.amazon.awssdk.services.secretsmanager.model.SecretsManagerException;
+// snippet-end:[secretsmanager.java2.get_secret.import]
+
+/**
+ * Before running this Java V2 code example, set up your development
+ * environment, including your credentials.
+ *
+ * For more information, see the following documentation topic:
+ *
+ * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
+ *
+ * We recommend that you cache your secret values by using client-side caching.
+ *
+ * Caching secrets improves speed and reduces your costs. For more information,
+ * see the following documentation topic:
+ *
+ * https://docs.aws.amazon.com/secretsmanager/latest/userguide/retrieving-secrets.html
+ */
+public class GetSecretValue {
+    public static void main(String[] args) {
+        final String usage = """
+
+                Usage:
+                    <secretName>\s
+
+                Where:
+                    secretName - The name of the secret (for example, tutorials/MyFirstSecret).\s
+                """;
+
+      //  if (args.length != 1) {
+       //     System.out.println(usage);
+      //      System.exit(1);
+      //  }
+
+        String secretName = "mysecret" ; //args[0];
+        Region region = Region.US_EAST_1;
+        SecretsManagerClient secretsClient = SecretsManagerClient.builder()
+                .region(region)
+                .build();
+
+        getValue(secretsClient, secretName);
+        secretsClient.close();
+    }
+
+    public static void getValue(SecretsManagerClient secretsClient, String secretName) {
+        try {
+            GetSecretValueRequest valueRequest = GetSecretValueRequest.builder()
+                    .secretId(secretName)
+                    .build();
+
+            GetSecretValueResponse valueResponse = secretsClient.getSecretValue(valueRequest);
+            String secret = valueResponse.secretString();
+            System.out.println(secret);
+
+        } catch (SecretsManagerException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+    }
+}
+// snippet-end:[secretsmanager.java2.get_secret.main]
